requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Gemini AI (FREE Alternative to OpenAI)
google-generativeai==0.3.2

# Database (using simpler versions with pre-built wheels)
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Kafka Event Streaming
aiokafka==0.10.0

# Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Twilio WhatsApp
twilio==8.11.0

# Utilities
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging
structlog==24.1.0

# Note: If installation fails on Windows, you may need:
# 1. Visual Studio Build Tools: https://visualstudio.microsoft.com/downloads/
# 2. Or use WSL2 (Windows Subsystem for Linux)
# 3. Or use pre-built Docker containers
